feat: modularizando os algoritimos em uma lib só
