--- bucketSort.h
@@ -1,5 +1,0 @@
-void insertionSort(int *array, int tamanho);
-
-void bucketSort(int *vetor, int tamanho);
-
-//void printArray(int arr[], int size);
