--- countingSort.h
@@ -1,1 +1,0 @@
-void countingSort(int inputArray[], int numElementos);
