+++ eda2_p1.depend
@@ -73,51 +73,35 @@
 1745017930 source:z:\my drive\projetos\p1-eda2\mergesort.c
 	"aux_p1.h"
 
-1745004966 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\aux_p1.c
-	<stdio.h>
-	<stdlib.h>
-	<locale.h>
-	<time.h>
-	<conio.h>
-	<time.h>
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\aux_p1.c
+	"aux_p1.h"
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\bucketsort.c
-	<stdio.h>
-	<stdlib.h>
-	"bucketSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\bucketsort.c
+	"aux_p1.h"
 
 1745003082 c:\users\eduardo\desktop\p1_eda2\p1-eda2\bucketsort.h
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\countingsort.c
-	<stdio.h>
-	<stdlib.h>
-	"countingSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\countingsort.c
+	"aux_p1.h"
 
 1745003082 c:\users\eduardo\desktop\p1_eda2\p1-eda2\countingsort.h
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\heapsort.c
-	<stdio.h>
-	"heapSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\heapsort.c
+	"aux_p1.h"
 
 1745003082 c:\users\eduardo\desktop\p1_eda2\p1-eda2\heapsort.h
 
-1745005065 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\main.c
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\main.c
+	"aux_p1.h"
+
+1745029371 c:\users\eduardo\desktop\p1_eda2\p1-eda2\aux_p1.h
 	<stdio.h>
 	<stdlib.h>
 	<locale.h>
 	<conio.h>
 	<time.h>
-	"aux_p1.h"
-	"radixSort.h"
-	"quickSort.h"
-	"shellSort.h"
-	"bucketSort.h"
-	"countingSort.h"
-	"heapSort.h"
-	"mergeSort.h"
-	"timSort.h"
-
-1745003082 c:\users\eduardo\desktop\p1_eda2\p1-eda2\aux_p1.h
+	<math.h>
+	"algoritimos.h"
 
 1745003082 c:\users\eduardo\desktop\p1_eda2\p1-eda2\radixsort.h
 
@@ -129,15 +113,11 @@
 
 1745003082 c:\users\eduardo\desktop\p1_eda2\p1-eda2\timsort.h
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\mergesort.c
-	<stdio.h>
-	<stdlib.h>
-	"mergeSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\mergesort.c
+	"aux_p1.h"
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\quicksort.c
-	<stdio.h>
-	<stdlib.h>
-	"quickSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\quicksort.c
+	"aux_p1.h"
 
 1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\radixsorte.c
 	<stdio.h>
@@ -146,15 +126,11 @@
 	<time.h>
 	"radixSort.h"
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\shellsort.c
-	<stdio.h>
-	<stdlib.h>
-	"shellSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\shellsort.c
+	"aux_p1.h"
 
-1745003082 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\timsort.c
-	<stdio.h>
-	<stdlib.h>
-	"timSort.h"
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\timsort.c
+	"aux_p1.h"
 
 1745007293 source:c:\projetos\ifsp\terceirosemestre\estruturasdedados02-ifsp\p1-eda2\aux_p1.c
 	<stdio.h>
@@ -228,4 +204,13 @@
 
 1745017930 source:z:\my drive\projetos\p1-eda2\radixsort.c
 	"aux_p1.h"
+
+1745020381 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\radixsort.c
+	"aux_p1.h"
+
+1745023035 source:c:\users\eduardo\desktop\p1_eda2\p1-eda2\algoritimos.c
+	<stdio.h>
+	<stdlib.h>
+
+1745022855 c:\users\eduardo\desktop\p1_eda2\p1-eda2\algoritimos.h
