--- heapSort.c
@@ -1,75 +1,0 @@
-//GU3008398 - Wesley Silva
-//GU301553X - Hugo Barbosa
-//GU3052834 - Luiz Fernando
-
-#include "aux_p1.h"
-
-void swap(int *a, int *b) {
-    int temp = *a;
-    *a = *b;
-    *b = temp;
-}
-
-// Função para criar um Max-Heap
-void maxHeapify(int arr[], int n, int i) {
-    int largest = i; // Inicializa o maior como o nó
-    int left = 2 * i + 1;
-    int right = 2 * i + 2;
-
-    if (left < n && arr[left] > arr[largest])
-        largest = left;
-
-    if (right < n && arr[right] > arr[largest])
-        largest = right;
-
-    if (largest != i) {
-        swap(&arr[i], &arr[largest]);
-        maxHeapify(arr, n, largest);
-    }
-}
-
-// Função para criar um Min-Heap
-void minHeapify(int arr[], int n, int i) {
-    int smallest = i;
-    int left = 2 * i + 1;
-    int right = 2 * i + 2;
-
-    if (left < n && arr[left] < arr[smallest])
-        smallest = left;
-
-    if (right < n && arr[right] < arr[smallest])
-        smallest = right;
-
-    if (smallest != i) {
-        swap(&arr[i], &arr[smallest]);
-        minHeapify(arr, n, smallest);
-    }
-}
-
-// Função HeapSort usando Max-Heap (ordem crescente)
-void heapSortAscending(int arr[], int n) {
-    for (int i = n / 2 - 1; i >= 0; i--)
-        maxHeapify(arr, n, i);
-
-    for (int i = n - 1; i > 0; i--) {
-        swap(&arr[0], &arr[i]);
-        maxHeapify(arr, i, 0);
-    }
-}
-
-// Função HeapSort usando Min-Heap (ordem decrescente)
-void heapSortDescending(int arr[], int n) {
-    for (int i = n / 2 - 1; i >= 0; i--)
-        minHeapify(arr, n, i);
-
-    for (int i = n - 1; i > 0; i--) {
-        swap(&arr[0], &arr[i]);
-        minHeapify(arr, i, 0);
-    }
-}
-
-/*void printArray(int arr[], int n) {
-    for (int i = 0; i < n; i++)
-        printf("%d ", arr[i]);
-    printf("\n");
-}*/
