--- heapSort.h
@@ -1,15 +1,0 @@
-void swap(int *a, int *b);
-
-// Função para criar um Max-Heap
-void maxHeapify(int arr[], int n, int i);
-
-// Função para criar um Min-Heap
-void minHeapify(int arr[], int n, int i);
-
-// Função HeapSort usando Max-Heap (ordem crescente)
-void heapSortAscending(int arr[], int n);
-
-// Função HeapSort usando Min-Heap (ordem decrescente)
-void heapSortDescending(int arr[], int n);
-
-//void printArray(int arr[], int n);
