--- mergeSort.c
@@ -1,52 +1,0 @@
-#include "aux_p1.h"
-
-
-void mergesort(int *v, int inicio, int fim){
-    int meio;
-    if(inicio < fim){
-        meio = (inicio + fim) / 2;
-        mergesort(v, inicio, meio);
-        mergesort(v, meio+1, fim);
-        merge(v, inicio, meio, fim);
-    }
-}
-
-void merge(int *v, int inicio, int meio, int fim){
-    int *temp, p1, p2, tamanho,i ,j, k;
-    int fim1 = 0, fim2 = 0;
-    tamanho = fim - inicio + 1; //tamanho do novo vetor
-    p1 = inicio;
-    p2 = meio+1;
-    temp = (int *) malloc(tamanho*sizeof(int));
-    if(temp != NULL){
-        for(i = 0; i<tamanho; i++){/*continua at'e preencher todos os indices do novo vetor*/
-            if(fim1 == 0 && fim2 == 0){/*veifica se algum dos subvetores chegou ao final*/
-                //------ordenando-------
-                if(v[p1] < v[p2]){
-                    temp[i] = v[p1];
-                    p1++;
-                }else{
-                    temp[i] = v[p2];
-                    p2++;
-                }
-                //-----verificando se o vetor acabou------
-                if(p1> meio) {fim1 = 1;}
-                if(p2> fim) {fim2 = 1;}
-
-            }else{
-                if(fim1 == 1){/*foi o p1 que acabou? Copia o resto do p1 no vetor*/
-                    temp[i] = v[p2];
-                    p2++;
-                }else{        /*foi o p2 que acabou? copia o resto do p1 no vetor*/
-                    temp[i] = v[p1];
-                    p1++;
-                }
-            }
-        }
-        //----copiando temp para o vetor passado como parametro;
-        for(j = 0, k= inicio; j<tamanho; j++, k++){
-            v[k] = temp[j];
-        }
-    }
-    free(temp);//liberar o vetor temporario
-}
