--- mergeSort.h
@@ -1,2 +1,0 @@
-void mergesort(int *v, int inicio, int fim);
-void merge(int *v, int inicio, int meio, int fim);
