--- shellSort.h
@@ -1,5 +1,0 @@
-void InsertionSort(int arr[], int tam);
-//ordenação de vetor de inteiros por Insertion Sort
-
-void ShellSort(int arr[], int tam);
-//ordenação de vetor de inteiros por Shell Sort
