--- timSort.h
@@ -1,5 +1,0 @@
-int min(int a, int b); // Função para definir o mínimo de dois números
-void insertSort(int array[], int esq, int dir); // Função para inserir o método InsertionSort
-void mergeSort(int array[], int x, int y, int z);
-void timSort(int array[], int n); // Função que executa o método TimSort:
-
